requirements.txt
pytube==15.0.0
moviepy==1.0.3
openai-whisper==20231117
ffmpeg==1.4
openai==1.30.1
streamlit==1.35.0
python-dotenv==1.0.1
